Implem new approach for f_urb estimation
Areas with impervious fraction of less than 0.5 are assigned to rural,
the rest is assigned to urban (with same weight). This was supported by
visual investigation of the dataset. This results in larger values of
fr_urb, more consistent with visual examination of the area. Moreover,
it allows selecting (with the same criteria) the street trees from other
urban vegetation (TO BE DONE AS NEXT STEP).